create CompareTradeHelpers
